Environment Reorganize (requires more testing)
