+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "python.analysis.autoFormatStrings": true,
     "python.analysis.autoImportCompletions": true,
-    "python.analysis.importFormat": "relative"
+    "python.analysis.importFormat": "relative",
+    "python.analysis.typeCheckingMode": "basic"
 }
