+++ ducklingscript/compiler/commands/start.py
@@ -2,7 +2,7 @@ from typing import Any
 from .bases.doc_command import ArgReqType
 
 from .bases.simple_command import Line, SimpleCommand
-from ducklingscript.compiler.environment import Environment
+from ducklingscript.compiler.environments.environment import Environment
 from ducklingscript.compiler.pre_line import PreLine
 from ducklingscript.compiler.stack_return import CompiledReturn
 from ..errors import (
