+++ ducklingscript/compiler/commands/while_loop.py
@@ -129,7 +129,7 @@ class While(BlockCommand):
 
             with self.stack.add_stack_above(code_block) as new_stack:
                 if var_name is not None:
-                    new_stack.env.new_var(var_name, count)
+                    new_stack.env.var.new_var(var_name, count)
 
                 if not Tokenizer.tokenize(argument, new_stack, new_stack.env):
                     break
