+++ ducklingscript/compiler/environments/__init__.py
@@ -1,0 +1,4 @@
+from .base_environment import BaseEnvironment
+from .environment import Environment
+from .project_environment import ProjectEnvironment
+from .variable_environment import VariableEnvironment
